190124-1718

+++ commands/adm-clear.js
@@ -6,6 +6,7 @@ module.exports.run = async (message, cmd, args) => {
     if (!args[0]) return message.channel.send("¯\\_(ツ)_/¯ *My crystal ball broke, please tell me, how many to clear?*")
     if (isNaN(args[0])) return message.channel.send("¯\\_(ツ)_/¯")
     let toclear = args[0]
+    if (toclear > 100) return message.channel.send("Value should be less than or equal to 100")
     await message.channel.bulkDelete(toclear + 1).then(() => {
         message.channel.send(`👍 Cleared **${toclear}** messages.`).then(msg => msg.delete(2500))
     })
